Refactor to pass CodeClimate
